Includes the images used, frontend part in html, css and js files and the backend part in python using flask
It has all the images required, the python file cards magic.py which is the backend part which has the logic for the trick and the serves the images.
It also has the index.html, script.js and styles.css which serves as the front end of the project and shows the trick.

Index.html is simply for the layout. The entire frontend is developed in javascript in the file script.js